Use fork instead of spawn for tns command
When using spawn we communicate to the tns process through stdin. This is causing exception in buildMetadata step when gradle wrapper is used.
For some reason it throws error in java code.
Replace the spawn with fork. Fork requires JavaScript module, so use the nativescript-cli.js from bin folder in case the provided options do not send path to tns.
The child process will send "ready" to the parent, when it is ready to read data. In this moment the parent will send the configuration and the unit testing will start.