Do not schedule execution in case `--watch` is not set
In case `--watch` is not set, CLI will set `singleRun` option of karma config to true. This way on browser register event, karma will automatically run the tests.
In case `--watch` is set, any change in code should start new tests, so when browser is registered, schedule new execution and karma will add the test execution to the queue and execute them when available.